change name of more files and refactor layout of more infomation form and specifications form
