+++ src/components/sellerPages/inventory/newItem/main_info.js
